last couple of tweaks
Next step could be to take, for instance, Section 3 and make it a new /Sec/Exclusions_v0.md.  That involves replacing it with:

3.=[Wx/com/stlstartuplawyer.com/NDA/Sec/Exclusions_v0.md]

and then (in Sec/Exclusions_v0.md) removing the leading 3. across all of the sections (the 3. is provided by prefixing).  We could then put in 

3.Sec={Exclude.Sec}

and then on mouseover each of the keys would read as Exclude.2.sec etc.   Can do that for all sections.  Can then change the outline to read <ol><li>{DefConfInfo.Sec}<li>{NotDisclose.Sec}<li>{Exclude.Sec}<li>....